@@ -1,93 +1,162 @@
 {
-  "name": "react-semantic.ui-starter",
-  "version": "1.0.1",
-  "main": "common/index.jsx",
-  "description": "React/Redux/Jest/Webpack/Semantic.UI boilerplate/example.",
+  "name": "suicrux",
+  "version": "3.0.1",
+  "main": "src/client/index.jsx",
+  "description": "Ultimate universal starter with code-splitting, server-side rendering and i18n",
   "dependencies": {
-    "lodash": "^4.17.4",
-    "prop-types": "^15.5.8",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-redux": "^5.0.4",
-    "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
-    "react-router-redux": "^5.0.0-alpha.6",
-    "redux": "^3.6.0",
+    "babel-polyfill": "^6.26.0",
+    "body-parser": "^1.18.2",
+    "chokidar": "^2.0.2",
+    "compression": "^1.7.2",
+    "cookie-parser": "^1.4.3",
+    "express": "^4.16.2",
+    "helmet": "^3.11.0",
+    "history": "^4.7.2",
+    "isomorphic-fetch": "^2.2.1",
+    "lodash": "^4.17.5",
+    "morgan": "^1.9.0",
+    "offline-plugin": "^4.9.0",
+    "raven": "^2.4.1",
+    "react": "^16.2.0",
+    "react-async-bootstrapper": "^2.1.0",
+    "react-async-component": "^1.0.2",
+    "react-dom": "^16.2.0",
+    "react-ga": "^2.4.1",
+    "react-headroom": "^2.2.2",
+    "react-helmet": "^5.2.0",
+    "react-intl": "^2.4.0",
+    "react-redux": "^5.0.7",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
+    "react-router-redux": "^5.0.0-alpha.8",
+    "react-tree-walker": "^4.0.2",
+    "redux": "^3.7.2",
+    "redux-promise-middleware": "^5.0.0",
     "redux-thunk": "^2.2.0",
-    "semantic-ui-css": "^2.2.10",
-    "semantic-ui-react": "^0.68.2",
-    "store2": "^2.5.1",
-    "whatwg-fetch": "^2.0.3"
+    "semantic-ui-css": "^2.2.14",
+    "semantic-ui-react": "^0.80.0",
+    "serialize-javascript": "^1.4.0",
+    "styled-components": "^3.1.6"
   },
   "devDependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.24.1",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^7.0.0",
-    "babel-plugin-lodash": "^3.2.11",
-    "babel-plugin-transform-react-jsx": "^6.24.1",
-    "babel-plugin-transform-regenerator": "^6.24.1",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-plugin-webpack-alias": "^2.1.2",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.24.1",
+    "@commitlint/cli": "^6.1.0",
+    "@commitlint/config-conventional": "^6.1.0",
+    "assets-webpack-plugin": "^3.5.1",
+    "autodll-webpack-plugin": "^0.3.9",
+    "autoprefixer": "^8.0.0",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.2.1",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-dynamic-import-node": "^1.2.0",
+    "babel-plugin-lodash": "^3.3.2",
+    "babel-plugin-module-resolver": "^3.1.0",
+    "babel-plugin-styled-components": "1.7.1",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-flow": "^6.23.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-react-optimize": "^1.0.1",
     "babel-preset-stage-0": "^6.24.1",
-    "babel-runtime": "^6.23.0",
-    "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.28.0",
-    "enzyme": "^2.8.2",
-    "eslint": "^3.19.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-loader": "^1.7.1",
-    "eslint-plugin-babel": "^4.1.1",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-node": "^4.2.2",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-react": "^6.10.3",
+    "babel-register": "6.26.0",
+    "babel-runtime": "^6.26.0",
+    "caniuse-lite": "^1.0.30000809",
+    "circular-dependency-plugin": "^4.4.0",
+    "compression-webpack-plugin": "^1.1.7",
+    "copy-webpack-plugin": "^4.4.1",
+    "cross-env": "^5.1.3",
+    "css-loader": "^0.28.9",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "enzyme-to-json": "^3.3.1",
+    "eslint": "^4.18.0",
+    "eslint-config-standard": "^11.0.0",
+    "eslint-loader": "^1.9.0",
+    "eslint-plugin-babel": "^4.1.2",
+    "eslint-plugin-flowtype": "^2.45.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-node": "^6.0.0",
+    "eslint-plugin-promise": "^3.6.0",
+    "eslint-plugin-react": "^7.6.1",
     "eslint-plugin-standard": "^3.0.1",
-    "extract-text-webpack-plugin": "^2.1.0",
-    "file-loader": "^0.11.1",
-    "friendly-errors-webpack-plugin": "^1.6.1",
-    "glob": "^7.1.1",
-    "html-webpack-plugin": "^2.28.0",
-    "img-loader": "^2.0.0",
-    "jest": "^19.0.2",
-    "json-server": "^0.10.0",
-    "minimist": "^1.2.0",
-    "node-sass": "^4.5.2",
-    "nodemon": "^1.11.0",
-    "offline-plugin": "^4.7.0",
-    "optimize-css-assets-webpack-plugin": "^1.3.1",
-    "preload-webpack-plugin": "^1.2.2",
-    "purifycss-webpack": "^0.6.1",
-    "react-addons-perf": "^15.4.2",
-    "react-addons-test-utils": "^15.5.1",
-    "react-hot-loader": "^3.0.0-beta.6",
-    "redux-devtools-extension": "^2.13.0",
-    "redux-mock-store": "^1.2.3",
-    "resolve-url-loader": "^2.0.2",
-    "sass-loader": "^6.0.3",
-    "style-loader": "^0.16.1",
-    "url-loader": "^0.5.8",
-    "webpack": "2.4.1",
-    "webpack-dev-middleware": "^1.10.2",
-    "webpack-hot-middleware": "^2.18.0",
-    "webpack-manifest-plugin": "^1.1.0",
-    "why-did-you-update": "0.0.8"
+    "extract-text-webpack-plugin": "^3.0.2",
+    "file-loader": "^1.1.7",
+    "flow-bin": "^0.69.0",
+    "husky": "^0.14.3",
+    "jest": "^22.3.0",
+    "jest-styled-components": "^5.0.0",
+    "nock": "^9.1.6",
+    "node-noop": "^1.0.0",
+    "node-sass": "^4.7.2",
+    "optimize-css-assets-webpack-plugin": "^3.2.0",
+    "optimize-js-plugin": "0.0.4",
+    "postcss": "^6.0.19",
+    "postcss-browser-reporter": "^0.5.0",
+    "postcss-cssnext": "^3.1.0",
+    "postcss-import": "^11.1.0",
+    "postcss-loader": "^2.1.0",
+    "postcss-reporter": "^5.0.0",
+    "razzle-dev-utils": "^0.8.10",
+    "react-hot-loader": "^3.1.3",
+    "react-test-renderer": "^16.2.0",
+    "redux-devtools-extension": "^2.13.2",
+    "redux-mock-store": "^1.5.1",
+    "resolve-url-loader": "^2.2.1",
+    "rimraf": "^2.6.2",
+    "sass-loader": "^6.0.6",
+    "source-map-support": "^0.5.3",
+    "standard-version": "^4.3.0",
+    "start-server-webpack-plugin": "^2.2.1",
+    "style-loader": "^0.20.2",
+    "stylelint": "^9.0.0",
+    "stylelint-config-standard": "^18.1.0",
+    "url-loader": "^0.6.2",
+    "webpack": "^3.11.0",
+    "webpack-assets-manifest": "^1.0.0",
+    "webpack-bundle-analyzer": "^2.10.0",
+    "webpack-dev-server": "^2.11.1",
+    "webpack-hot-middleware": "^2.24.2",
+    "webpack-manifest-plugin": "^1.3.2",
+    "webpack-node-externals": "^1.6.0",
+    "write-file-webpack-plugin": "^4.2.0"
+  },
+  "optionalDependencies": {
+    "immutable": "^3.8.2",
+    "normalizr": "^3.2.4",
+    "query-string": "^6.0.0",
+    "reselect": "^3.0.1",
+    "store2": "^2.7.0",
+    "js-cookie": "^2.2.0",
+    "sinon": "^5.0.0",
+    "chalk": "^2.3.1"
   },
   "scripts": {
-    "db": "nodemon ./db",
-    "build": "webpack --config webpack_config/webpack.prod.js",
-    "build:demo": "BUILD_DEMO=true webpack --config webpack_config/webpack.prod.js",
-    "test": "jest --config=jest_config/jest.config.json --coverage --forceExit || true",
-    "dev": "node webpack_config/server.js",
-    "ssr": "nodemon --exec REACT_WEBPACK_ENV='dev' babel-node ./common/server"
+    "build": "npm run frontend_build && npm run server_build",
+    "dev": "cross-env NODE_ENV=development babel-node ./webpack_config/devServer.js",
+    "server_build": "cross-env NODE_ENV=production webpack --config webpack_config/server/webpack.prod.babel.js",
+    "frontend_build": "cross-env NODE_ENV=production webpack --config webpack_config/client/webpack.prod.babel.js",
+    "test": "cross-env NODE_ENV=test jest --config=jest_config/jest.config.json --coverage --forceExit || true",
+    "snapshot_update": "cross-env NODE_ENV=test jest --config=jest_config/jest.config.json --coverage --updateSnapshot",
+    "start": "node ./dist/server",
+    "lint:styles": "stylelint ./src/common/**/*.scss -s scss --config ./.stylelintrc --fix",
+    "lint:scripts": "eslint -c ./.eslintrc.json ./src ./webpack_config --ignore-pattern dist",
+    "flow": "flow",
+    "release": "standard-version",
+    "prepush": "npm run lint:scripts",
+    "commitmsg": "commitlint -e $GIT_PARAMS"
   },
-  "author": "Vladimir Metnew <vladimirmetnew@gmail.com>",
+  "engines": {
+    "node": ">=8"
+  },
+  "author": {
+    "name": "Vladimir Metnew",
+    "email": "vladimirmetnew@gmail.com"
+  },
   "license": "MIT",
-  "repository": "https://github.com/Metnew/react-semantic.ui-starter.git",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Metnew/suicrux.git"
+  },
   "keywords": [
     "react",
     "redux",
